+++ tools/globals.js
@@ -89,6 +89,7 @@ GetMaxBattlefieldID GetBattlefieldStatus AcceptBattlefieldPort
 GetBattlefieldWinner LeaveBattlefield GetBattlegroundInfo JoinBattlefield
 GetRandomDungeonBestChoice SetLFGDungeon JoinLFG
 GetInstanceInfo LeaveParty
+GetCombatRating GetCombatRatingBonus
 strtrim strsplit strjoin strlower strupper strsub strlen strfind strmatch strrep
 tinsert tremove twipe wipe format gsub gmatch max min abs floor ceil
 date time difftime random